# src/routes/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Darve Server API",
    "description": "API documentation for the Darve Server including users, posts, notifications, auth, discussions, tasks, wallet, and follows",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Development server"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/users/current/password": {
      "post": {
        "tags": ["Users"],
        "summary": "Set user password",
        "description": "Set password for current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordInput"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password set successfully"
          },
          "400": {
            "description": "Invalid password"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update user password",
        "description": "Update password for current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordInput"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Password updated successfully."
                }
              }
            }
          },
          "400": {
            "description": "Invalid password"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/users/current/following/posts": {
      "get": {
        "tags": ["Users"],
        "summary": "Get posts from followed users",
        "description": "Get posts from users that the current user follows",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Posts from followed users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscussionPostView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/current/email/verification/start": {
      "post": {
        "tags": ["Users"],
        "summary": "Start email verification",
        "description": "Start email verification process",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationStartInput"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationStartInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verification started"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/current/email/verification/confirm": {
      "post": {
        "tags": ["Users"],
        "summary": "Confirm email verification",
        "description": "Confirm email verification with code",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationConfirmInput"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationConfirmInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/current": {
      "get": {
        "tags": ["Users"],
        "summary": "Get current user",
        "description": "Get current authenticated user information",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalUser"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Update current user",
        "description": "Update current user profile",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProfileSettingsFormInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalUser"
                }
              }
            }
          },
          "400": {
            "description": "Validation error - invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/current/posts": {
      "post": {
        "tags": ["Users"],
        "summary": "Create post for current user",
        "description": "Create a new post for the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PostInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "summary": "Get current user posts",
        "description": "Get posts created by the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "topic_id",
            "in": "query",
            "description": "Filter by topic ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Pagination start index",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "Number of posts to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/current/otp/enable": {
      "post": {
        "tags": ["Users"],
        "summary": "Enable OTP for current user",
        "description": "Enable two-factor authentication (OTP) for the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OTP enabled successfully, returns QR code URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "QR code URL for OTP setup"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/current/otp/disable": {
      "post": {
        "tags": ["Users"],
        "summary": "Disable OTP for current user",
        "description": "Disable two-factor authentication (OTP) for the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OTP disabled successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/current/otp/validate": {
      "post": {
        "tags": ["Users"],
        "summary": "Validate OTP token",
        "description": "Validate OTP token and get authenticated session",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpValidationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP validated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "403": {
            "description": "Invalid OTP token"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Search users",
        "description": "Search for users",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalUser"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResetPasswordInput": {
        "type": "object",
        "required": ["password", "code", "email_or_username"],
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "New password (minimum 6 characters)"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "description": "Verification code"
          },
          "email_or_username": {
            "type": "string",
            "description": "Email address or username"
          }
        }
      },
      "SetPasswordInput": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "Password (minimum 6 characters)"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT authentication token"
          },
          "user": {
            "$ref": "#/components/schemas/LocalUser"
          }
        }
      },
      "LocalUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "full_name": {
            "type": "string",
            "nullable": true,
            "description": "Full name"
          },
          "birth_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Birth date"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "Phone number"
          },
          "email_verified": {
            "type": "string",
            "nullable": true,
            "description": "Verified email address"
          },
          "bio": {
            "type": "string",
            "nullable": true,
            "description": "User biography"
          },
          "social_links": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "Social media links"
          },
          "image_uri": {
            "type": "string",
            "nullable": true,
            "description": "Profile image URI"
          },
          "has_otp": {
            "type": "boolean",
            "nullable": true,
            "description": "Whether user has OTP enabled"
          }
        }
      },
...
      "EmailVerificationStartInput": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address to verify"
          }
        }
      },
      "EmailVerificationConfirmInput": {
        "type": "object",
        "required": ["email", "code"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "description": "6-digit verification code"
          }
        }
      },
      "ProfileSettingsFormInput": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true,
            "minLength": 6,
            "pattern": "^[A-Za-z0-9_]{6,}$",
            "description": "Username (optional) - Letters, numbers and '_'. Minimum 6 characters"
          },
          "full_name": {
            "type": "string",
            "minLength": 6,
            "nullable": true,
            "description": "Full name (minimum 6 characters, optional)"
          },
          "birth_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Birth date in RFC3339 format (optional) - Must be between 10 and 120 years ago"
          },
          "image_url": {
            "type": "string",
            "format": "binary",
            "nullable": true,
            "description": "Profile image file (optional)"
          },
          "social_links": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "nullable": true,
            "description": "Social media links (optional) - Must be HTTPS URLs from x.com, instagram.com, youtube.com, or facebook.com. No duplicate domains allowed."
          }
        }
      },
      "PostInput": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "description": "Post title (minimum 5 characters)"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "description": "Post content (minimum 1 characters)"
          },
          "topic_id": {
            "type": "string",
            "description": "Topic ID for the post"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "description": "Tags for the post (maximum 5)"
          },
          "file_1": {
            "type": "string",
            "format": "binary",
            "description": "Optional file attachment"
          }
        }
      },
...
      "OtpValidationInput": {
        "type": "object",
        "required": ["token"],
        "properties": {
          "token": {
            "type": "string",
            "description": "OTP token from authenticator app"
          }
        }
      },
...
      "DiscussionPostView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Post ID"
          },
          "created_by_name": {
            "type": "string",
            "description": "Username of post creator"
          },
          "title": {
            "type": "string",
            "description": "Post title"
          },
          "content": {
            "type": "string",
            "description": "Post content"
          },
          "discussion_id": {
            "type": "string",
            "description": "Discussion ID"
          },
          "r_created": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    }
  },
...
}